DOMINATOR_V3_REPLIT/requirements.txt
python-dotenv
requests
python-telegram-bot==13.15
solana==0.36.6
solders==0.26.0
json5
matplotlib
telethon
